move env to root
