+++ app.js
@@ -2,7 +2,7 @@ const express = require("express");
 const path = require("path");
 const cookieParser = require("cookie-parser");
 const logger = require("morgan");
-const connectDB = require("./config/db");
+const connectDB = require("./db");
 
 const dotenv = require("dotenv");
 dotenv.config();
